--- example.json
@@ -1,8 +1,0 @@
-{
-  "users": {
-      "student_id": {
-        "Table": "students",
-        "Column": "student_id"
-      }
-  }
-}
