+++ go.mod
@@ -3,6 +3,18 @@ module github.com/Keith1039/Capstone_Test
 go 1.23.3
 
 require (
-	github.com/jimsmart/schema v0.2.1 // indirect
-	github.com/lib/pq v1.10.9 // indirect
+	gioui.org v0.7.1
+	github.com/jimsmart/schema v0.2.1
+	github.com/lib/pq v1.10.9
+)
+
+require (
+	gioui.org/cpu v0.0.0-20210817075930-8d6a761490d2 // indirect
+	gioui.org/shader v1.0.8 // indirect
+	github.com/go-text/typesetting v0.1.1 // indirect
+	golang.org/x/exp v0.0.0-20240707233637-46b078467d37 // indirect
+	golang.org/x/exp/shiny v0.0.0-20240707233637-46b078467d37 // indirect
+	golang.org/x/image v0.18.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	golang.org/x/text v0.16.0 // indirect
 )
